+++ src/App.vue
@@ -38,6 +38,7 @@ import { useSwap } from '@/composables/useSwap'
 import { useAnalyticsStore } from '@/stores/analyticsStore'
 import { analytics } from '@/analytics'
 import { useRewardsStore } from '@/stores/rewardsStore'
+import Intercom from '@intercom/messenger-js-sdk'
 
 const dialogStore = useDialogStore()
 const { isAreaHidden } = storeToRefs(dialogStore)
@@ -182,5 +183,14 @@ onMounted(() => {
     addProvider(provider)
   })
   initSwapper()
+  console.log('INTERCOME_TEST', configs.INTERCOM_APP_ID)
+  if (configs.INTERCOM_APP_ID) {
+    Intercom({
+      app_id: configs.INTERCOM_APP_ID,
+      hide_default_launcher: false,
+      horizontal_padding: 90, // distance from side
+      vertical_padding: 20, // distance from bottom
+    })
+  }
 })
 </script>